% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_counts_to_rkpm.R
\name{ensembl_counts_to_rkpm}
\alias{ensembl_counts_to_rkpm}
\title{ensembl_counts_to_rkpm}
\usage{
ensembl_counts_to_rkpm(
  lengths_table_path = get_human_ensembl_to_hgnc_entrez_path(),
  file_prefix = "",
  input_file_path,
  output_dir,
  this_script_path = ""
)
}
\arguments{
\item{lengths_table_path}{Path to table for with cdna_length and transcript_id columns.}

\item{file_prefix}{Output file will have this prefix string appended to it.}

\item{input_file_path}{String path to input file. File should have samples by row, first columns should be the sample id with all subsequent columns headers in the Ensembl ENST format (no version).}

\item{output_dir}{Path to the output folder.}

\item{this_script_path}{Path to script that runs this function for documentation purposes}
}
\value{
A vector of paths to the output file.
}
\description{
Converts an Ensemble counts matrix into RKPM. <rant> RKPM is a terrible way to 
look at your data.  Only do it if you have to. If you choose cDNA length in 
your calculation then you loose all the non-coding transcripts.  If you use 
the full transcript length then you should account for the number of full and 
cds sequences which there's no way to know...</rant>
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
ensembl_counts_to_rkpm
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
