% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_counts_to_tpm.R
\name{ensembl_counts_to_tpm}
\alias{ensembl_counts_to_tpm}
\title{ensembl_counts_to_tpm}
\usage{
ensembl_counts_to_tpm(
  my_dt,
  lengths_table_path = get_human_ensembl_to_hgnc_entrez_path(),
  readme_path = NULL,
  this_script_path = ""
)
}
\arguments{
\item{my_dt}{data.table with samples by row, first column should be the sample id with all subsequent columns headers in the Ensembl ENST format (no version).}

\item{lengths_table_path}{Path to table for with cdna_length and transcript_id columns.}

\item{readme_path}{Optional path to which the comments will be appended.}

\item{this_script_path}{Path to script that runs this function for documentation purposes}
}
\value{
A vector of paths to the output file.
}
\description{
Converts an Ensemble ENST counts matrix into TPM.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
ensembl_counts_to_tpm
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
