% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_navigation.R
\name{get_human_ensembl_to_hgnc_entrez_path}
\alias{get_human_ensembl_to_hgnc_entrez_path}
\title{get_human_ensembl_to_hgnc_entrez_path}
\usage{
get_human_ensembl_to_hgnc_entrez_path()
}
\arguments{
\item{none}{}
}
\value{
A path to the table for converting Ensembl ENSTs to HGNC and Entrez IDs
}
\description{
get_human_ensembl_to_hgnc_entrez_path
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
