% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantsf_to_dt.R
\name{quantsf_to_dt}
\alias{quantsf_to_dt}
\title{quantsf_to_dt}
\usage{
quantsf_to_dt(
  input_file_paths,
  counts_or_tpm = "counts",
  readme_path = NULL,
  sample_key = "Run_ID",
  this_script_path = ""
)
}
\arguments{
\item{input_file_paths}{Named character vector of paths to the pipeline output data. Samples will be named by the names of the vector. If not named the will be named after the quant file.}

\item{counts_or_tpm}{'counts' or 'tpm'}

\item{readme_path}{Optional path to which the comments will be appended.}

\item{sample_key}{String to name the sample identifier column.}

\item{this_script_path}{Path to script that runs this function for documentation purposes}
}
\value{
A vector of paths to the output file.
}
\description{
Joins individual sample files into one tsv file.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
quantsf_to_dt
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
