% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantsf_to_matrix.R
\name{quantsf_to_matrix}
\alias{quantsf_to_matrix}
\title{quantsf_to_matrix}
\usage{
quantsf_to_matrix(
  counts_or_tpm = "counts",
  file_prefix = "",
  input_file_paths,
  output_dir,
  sample_key = "Run_ID",
  this_script_path = ""
)
}
\arguments{
\item{counts_or_tpm}{'counts' or 'tpm'}

\item{file_prefix}{Output files will have this prefix string appended to them}

\item{input_file_paths}{Named character vector of paths to the pipeline output data. Samples will be named by the names of the vector. If not named the will be named after the quant file.}

\item{output_dir}{Path to the output folder.}

\item{sample_key}{String to name the sample identifier column.}

\item{this_script_path}{Path to script that runs this function for documentation purposes}
}
\value{
A vector of paths to the output file.
}
\description{
Joins individual sample files into one tsv file.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
quantsf_to_matrix
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
