% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_counts_to_rkpm.R
\name{ensembl_counts_to_rkpm}
\alias{ensembl_counts_to_rkpm}
\title{ensembl_counts_to_rkpm}
\usage{
ensembl_counts_to_rkpm(
  my_dt,
  lengths_table_path = get_human_ensembl_to_hgnc_entrez_path(),
  readme_path = NULL,
  this_script_path = ""
)
}
\arguments{
\item{my_dt}{data.table with samples by row, first column should be the sample id with all subsequent columns headers in the Ensembl ENST format (no version).}

\item{lengths_table_path}{Path to table for with cdna_length and transcript_id columns.}

\item{readme_path}{Optional path to which the comments will be appended.}

\item{this_script_path}{Path to script that runs this function for documentation purposes}
}
\value{
A vector of paths to the output file.
}
\description{
Converts an Ensemble counts matrix into RKPM. <rant> RKPM is a terrible way to 
look at your data.  Only do it if you have to. If you choose cDNA length in 
your calculation then you loose all the non-coding transcripts.  If you use 
the full transcript length then you should account for the number of full and 
cds sequences which there's no way to know...</rant>
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
ensembl_counts_to_rkpm
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
