% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_counts_to_rkpm.R, R/quantsf_to_matrix.R
\name{quantsf_to_matrix}
\alias{quantsf_to_matrix}
\title{ensembl_counts_to_rkpm}
\usage{
quantsf_to_matrix(
  counts_or_tpm = "counts",
  file_prefix = "",
  input_file_paths,
  output_dir,
  sample_key = "Run_ID",
  this_script_path = ""
)

quantsf_to_matrix(
  counts_or_tpm = "counts",
  file_prefix = "",
  input_file_paths,
  output_dir,
  sample_key = "Run_ID",
  this_script_path = ""
)
}
\arguments{
\item{counts_or_tpm}{'counts' or 'tpm'}

\item{file_prefix}{Output files will have this prefix string appended to them}

\item{input_file_paths}{Named character vector of paths to the pipeline output data. Samples will be named by the names of the vector. If not named the will be named after the quant file.}

\item{output_dir}{Path to the output folder.}

\item{sample_key}{String to name the sample identifier column.}

\item{this_script_path}{Path to script that runs this function for documentation purposes}

\item{input_file_path}{String path to input file. File should have samples by row, first columns should be the sample id with all subsequent columns headers in the Ensembl ENST format (no version).}

\item{lengths_table_path}{Path to table for with cdna_length and transcript_id columns.}
}
\value{
A vector of paths to the output file.

A vector of paths to the output file.
}
\description{
Converts an Ensemble counts matrix into RKPM. <rant> RKPM is a terrible way to 
look at your data.  Only do it if you have to. If you choose cDNA length in 
your calculation then you loose all the non-coding transcripts.  If you use 
the full transcript length then you should account for the number of full and 
cds sequences which there's no way to know...</rant>

Joins individual sample files into one tsv file.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
ensembl_counts_to_rkpm
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
quantsf_to_matrix
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
